ML/machinelearning/requirements.txt
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.0
APScheduler==3.10.4
django-apscheduler==0.6.2
tensorflow==2.15.0
scikit-learn==1.4.1.post1
requests==2.31.0
mysqlclient==2.2.4
scikit-image==0.22.0
opencv-python